Refactored the algorithm.py to a separate subpackage
